fix failing tests hopefully Also split test/repl/helpers into multiple files
