+++ src/test/repl/helpers/ctx-repl.ts
@@ -1,8 +1,9 @@
-import { PassThrough } from 'stream';
-import { delay, TEST_DIR, tsNodeTypes, ctxTsNode } from '../helpers';
-import type { ExecutionContext } from 'ava';
-import { test, expect } from '../testlib';
+import type { ExecutionContext } from '@cspotcode/ava-lib';
 import { expectStream } from '@cspotcode/expect-stream';
+import { PassThrough } from 'stream';
+import type { ctxTsNode } from '../../helpers/ctx-ts-node';
+import { delay, tsNodeTypes } from '../../helpers/misc';
+import { TEST_DIR } from '../../helpers/paths';
 
 export interface CreateReplViaApiOptions {
   registerHooks: boolean;
@@ -96,47 +97,4 @@ export async function ctxRepl(t: ctxTsNode.T) {
       stderr: await stderrPromise,
     };
   }
-}
-
-export const macroReplNoErrorsAndStdoutContains = test.macro(
-  (script: string, contains: string, options?: Partial<ExecuteInReplOptions>) =>
-    async (t: ctxRepl.T) => {
-      macroReplInternal(t, script, contains, undefined, contains, options);
-    }
-);
-export const macroReplStderrContains = test.macro(
-  (
-      script: string,
-      errorContains: string,
-      options?: Partial<ExecuteInReplOptions>
-    ) =>
-    async (t: ctxRepl.T) => {
-      macroReplInternal(
-        t,
-        script,
-        undefined,
-        errorContains,
-        errorContains,
-        options
-      );
-    }
-);
-
-async function macroReplInternal(
-  t: ctxRepl.T,
-  script: string,
-  stdoutContains: string | undefined,
-  stderrContains: string | undefined,
-  waitPattern: string,
-  options?: Partial<ExecuteInReplOptions>
-) {
-  const r = await t.context.executeInRepl(script, {
-    registerHooks: true,
-    startInternalOptions: { useGlobal: false },
-    waitPattern,
-    ...options,
-  });
-  if (stderrContains) expect(r.stderr).toContain(stderrContains);
-  else expect(r.stderr).toBe('');
-  if (stdoutContains) expect(r.stdout).toContain(stdoutContains);
 }
