+++ src/test/repl/helpers/misc.ts
@@ -1,0 +1,3 @@
+import { tsNodeTypes, tsSupportsMtsCtsExtensions } from '../../helpers';
+
+export const replFile = tsSupportsMtsCtsExtensions ? '<repl>.cts' : '<repl>.ts';
